feat: distributed Implementation of the Parseable Server (#656)
This PR contains the ingester side changes
